Move kernel creation to separate file for later reuse
